backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.5
python-dotenv==1.2.1
yt-dlp==2026.3.17
